backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart